api_relation - DB수집시 글자수 제한 오류로 인해 max_length 값 증가.
