+++ django_app/openapi/urls.py
@@ -31,5 +31,4 @@ urlpatterns = [
     url(r'^area', snippet_area),
     # 분야별 검색 (search= genre)
     url(r'^genre', snippet_genre),
-
 ]
